+++ taskmonitor/__init__.py
@@ -1,4 +1,4 @@
 default_app_config = "taskmonitor.apps.TaskMonitorConfig"
 
-__version__ = "0.14.1a2"
+__version__ = "0.15.0"
 __title__ = "Task Monitor"